Remove tagging from list
This is something you can handle yourself when defining item views. The
item model must have an ID, meaning you have all the information you
need to wrap and tag components yourself. Better to do the simple thing
here and just pass down send, allowing the item to handle this wiring
itself if it needs to.